🔨 refactor(none): desktop & docs minimum code, refactor code
