+++ .github/workflows/codeql.yml
@@ -36,7 +36,7 @@ jobs:
       - name: Checkout repository
         uses: actions/checkout@v3
 
-      # Initializes the CodeQL tools for scanning.
+      # Initializes the CodeQL for scanning.
       - name: Initialize CodeQL
         uses: github/codeql-action/init@v2
         with:
