+++ apps/desktop/README.md
@@ -1,0 +1,5 @@
+# @templ/desktop
+
+TEMPL desktop module.
+
+[GitHub](https://github.com/rjoydip/templ/tree/main/packages/desktop)
