+++ apps/desktop/vitest.config.mts
@@ -1,0 +1,7 @@
+import { defineProject, mergeConfig } from 'vitest/config'
+import configShared from '../../config/vitest.shared.js'
+
+export default mergeConfig(
+  configShared,
+  defineProject({}),
+)
