+++ apps/docs/README.md
@@ -1,0 +1,5 @@
+# @templ/docs
+
+TEMPL docs module.
+
+[GitHub](https://github.com/rjoydip/templ/tree/main/packages/docs)
