+++ config/tsconfig/README.md
@@ -1,0 +1,5 @@
+# @templ/tsconfig
+
+TSconfig.
+
+[GitHub](https://github.com/rjoydip/templ/tree/main/config/tsconfig)
