+++ config/tsconfig/base.json
