+++ config/tsconfig/package.json
@@ -6,11 +6,11 @@
   "packageManager": "pnpm@8.15.1",
   "description": "TSconfig",
   "license": "MIT",
-  "homepage": "https://github.com/rjoydip/templ/tree/main/tools/tsconfig#readme",
+  "homepage": "https://github.com/rjoydip/templ/tree/main/config/tsconfig#readme",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/rjoydip/templ.git",
-    "directory": "tools/tsconfig"
+    "directory": "config/tsconfig"
   },
   "bugs": {
     "url": "https://github.com/rjoydip/templ/issues"
