+++ config/vitest.shared.js
@@ -1,6 +1,6 @@
 const configShared = {
   test: {
-    include: ['test/**/*.test.ts'],
+    include: ['{test,tests}/**/*.test.{ts,js}'],
     coverage: {
       enabled: true,
       reporter: ['text', 'json', 'html'],
