+++ generator/README.md
@@ -1,0 +1,5 @@
+# @templ/generator
+
+Generator module.
+
+[GitHub](https://github.com/rjoydip/templ/tree/main/generator)
