+++ generator/package.json
@@ -4,13 +4,13 @@
   "version": "0.0.1",
   "private": true,
   "packageManager": "pnpm@8.15.1",
-  "description": "Generator tools",
+  "description": "Generator",
   "license": "MIT",
-  "homepage": "https://github.com/rjoydip/templ/tree/main/tools/generator#readme",
+  "homepage": "https://github.com/rjoydip/templ/tree/main/generator#readme",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/rjoydip/templ.git",
-    "directory": "tools/generator"
+    "directory": "generator"
   },
   "bugs": {
     "url": "https://github.com/rjoydip/templ/issues"
