+++ generator/templates/basic-js/README.md
