+++ generator/templates/basic-js/package.json
