+++ generator/templates/basic-js/src/index.js
