+++ generator/templates/basic-js/test/index.test.js
