+++ generator/templates/basic-js/vitest.config.mjs
@@ -1,5 +1,5 @@
 import { defineProject, mergeConfig } from 'vitest/config'
-import configShared from '../../../../config/vitest.shared.js'
+import configShared from '../../../config/vitest.shared.js'
 
 export default mergeConfig(
   configShared,
