+++ generator/templates/basic-ts/esbuild.config.ts
@@ -1,0 +1,13 @@
+import { argv } from 'node:process'
+import * as esbuild from 'esbuild'
+import { getBuildConfig } from '@templ/config'
+
+const config = getBuildConfig()
+
+if (argv.slice(2).includes('--watch')) {
+  const ctx = await esbuild.context(config)
+  await ctx.watch()
+}
+else {
+  await esbuild.build(config)
+}
