+++ generator/templates/basic-ts/src/index.ts
