--- tools/generator/README.md
@@ -1,5 +1,0 @@
-# @templ/generator
-
-Generator module.
-
-[GitHub](https://github.com/rjoydip/templ/tree/main/tools/generator)
