--- tools/generator/templates/basic-ts/tsconfig.json
@@ -1,4 +1,0 @@
-{
-  "extends": "@templ/tsconfig/base.json",
-  "include": ["src/**/*.ts", "esbuild.config.ts"]
-}
