--- tools/tsconfig/README.md
@@ -1,5 +1,0 @@
-# @templ/tsconfig
-
-TSconfig.
-
-[GitHub](https://github.com/rjoydip/templ/tree/main/tools/tsconfig)
